+++ config/defaults.conf
@@ -400,11 +400,14 @@ qqp-alt_pair_attn = 1
 qqp-alt_val_interval = 1000
 qqp-alt_lr = 0.0003
 
-chunk_size = 10 //chunk size for the ON-LSTM encoder
-dropconnect = 0.5
-dropouti  = 0.3
-dropouth  = 0.3
-tying = 0 //LM tying of weights
+onlstm_chunk_size = 10  //Chunk downsizing factor for ON-LSTM master gate, dimensions
+                        //of master gate: D/C where C is the chunk downsizing factor.
+onlstm_dropconnect = 0.5  //Linear dropout between ON-LSTM layers.
+onlstm_dropouti  = 0.3 // Locked Dropout on input embeddings.
+onlstm_dropouth  = 0.3 // Locked Dropout between ON-LSTM layers.
+onlstm_tying = 0 // Language Modeling tying of weights.
+
+
 embeddings_train  = 0  // if set to 1, embeddings will be fine tuned.
 
 nli-prob {
