+++ src/tasks/lm.py
@@ -152,193 +152,3 @@ class WikiText103LMTask(WikiTextLMTask):
         self.files_by_split = {'train': os.path.join(path, "train.sentences.txt"),
                                'val': os.path.join(path, "valid.sentences.txt"),
                                'test': os.path.join(path, "test.sentences.txt")}
-
-@register_task('wsj', rel_path='Penn/')
-class WSJLanguageModelling(LanguageModelingTask):
-    """ Language modeling on a PTB dataset
-    See base class: LanguageModelingTask
-    """
-    def count_examples(self):
-        """Computes number of samples
-        Assuming every line is one example.
-        """
-        example_counts = {}
-        for split, split_path in self.files_by_split.items():
-            arr = [line.strip().split()+["<EOS>"] for line in open(split_path)]
-            allf = 0
-            for x in arr:
-                allf += len(x)
-            example_counts[split] = int(math.ceil(allf/self.max_seq_len))
-        self.example_counts = example_counts
-
-    
-    def process_split(self, split, indexers) -> Iterable[Type[Instance]]:
-        """Process a language modeling split by indexing and creating fields.
-        Args:
-            split: (list) a single list of sentences
-            indexers: (Indexer object) indexer to index input words
-        """
-        def _make_instance(sent):
-            ''' Forward targs adds <s> as a target for input </s>
-            and bwd targs adds </s> as a target for input <s>
-            to avoid issues with needing to strip extra tokens
-            in the input for each direction '''
-            d = {}
-            #import pdb;pdb.set_trace()
-            d["input"] = sentence_to_text_field(sent[:-1], indexers)
-            d["targs"] = sentence_to_text_field(sent[1:], self.target_indexer)
-            d["targs_b"] = sentence_to_text_field([sent[-1]] + sent[:-2], self.target_indexer)
-            return Instance(d)
-        for sent in split:
-            yield _make_instance(sent)
-    def load_data(self, path):
-        """Loading data file and tokenizing the text
-        Args:
-            path: (str) data file path
-        """
-        seq_len = self.max_seq_len
-        tokens = []
-        with open(path) as txt_fh:
-            for row in txt_fh:
-                toks = row.strip()
-                if not toks:
-                    continue
-                toks_v = toks.split()
-                toks = toks.split()+["<EOS>"]
-                tokens += toks
-            for i in range(0, len(tokens), seq_len):
-                yield tokens[i:i+seq_len]
-
-@register_task('toronto_lm', rel_path='toronto/')
-class TorontoLanguageModelling(LanguageModelingTask):
-    """ Language modeling on a PTB dataset
-    See base class: LanguageModelingTask
-    """
-    def count_examples(self):
-        """Computes number of samples
-        Assuming every line is one example.
-        """
-        example_counts = {}
-        for split, split_path in self.files_by_split.items():
-            arr = [line.strip().split()+["<EOS>"] for line in open(split_path)]
-            allf = 0
-            for x in arr:
-                allf += len(x)
-            example_counts[split] = int(math.ceil(allf/self.max_seq_len))
-        self.example_counts = example_counts
-
-    
-    def process_split(self, split, indexers) -> Iterable[Type[Instance]]:
-        """Process a language modeling split by indexing and creating fields.
-        Args:
-            split: (list) a single list of sentences
-            indexers: (Indexer object) indexer to index input words
-        """
-        def _make_instance(sent):
-            ''' Forward targs adds <s> as a target for input </s>
-            and bwd targs adds </s> as a target for input <s>
-            to avoid issues with needing to strip extra tokens
-            in the input for each direction '''
-            d = {}
-            #import pdb;pdb.set_trace()
-            d["input"] = sentence_to_text_field(sent[:-1], indexers)
-            d["targs"] = sentence_to_text_field(sent[1:], self.target_indexer)
-            d["targs_b"] = sentence_to_text_field([sent[-1]] + sent[:-2], self.target_indexer)
-            return Instance(d)
-        for sent in split:
-            yield _make_instance(sent)
-    def load_data(self, path):
-        """Loading data file and tokenizing the text
-        Args:
-            path: (str) data file path
-        """
-        seq_len = self.max_seq_len
-        tokens = []
-        with open(path) as txt_fh:
-            for row in txt_fh:
-                toks = row.strip()
-                if not toks:
-                    continue
-                toks_v = toks.split()
-                toks = toks.split()+["<EOS>"]
-                tokens += toks
-            for i in range(0, len(tokens), seq_len):
-                yield tokens[i:i+seq_len]
-
-
-
-@register_task('mnli_lm', rel_path='MNLI/')
-class MNLILanguageMdelling(LanguageModelingTask):
-    """ Language modeling on the MNLI dataset
-    See base class: LanguageModelingTask
-    """
-
-    def __init__(self, path, max_seq_len, name="mnli_lm", **kw):
-        """Init class
-        Args:
-            path: (str) path that the data files are stored
-            max_seq_len: (int) maximum length of one sequence
-            name: (str) task name
-        """
-        super().__init__(path, max_seq_len, name, **kw)
-        self.scorer1 = Average()
-        self.scorer2 = None
-        self.val_metric = "%s_perplexity" % self.name
-        self.val_metric_decreases = True
-        self.max_seq_len = max_seq_len
-        self.min_seq_len = 0
-        self.target_indexer = {"words": SingleIdTokenIndexer(namespace="tokens")}
-        self.files_by_split = {'train': os.path.join(path, "train.tsv"),
-                               'val': os.path.join(path, "dev_matched.tsv"),
-                               'test': os.path.join(path, "test_matched.tsv")}
-    def count_examples(self):
-        """Computes number of samples
-        Assuming every line is one example.
-        """
-        example_counts = {}
-        for split, split_path in self.files_by_split.items():
-            arr = [line.strip().split()+["<EOS>"] for line in open(split_path)]
-            allf = 0
-            for x in arr:
-                allf += len(x)
-            example_counts[split] = int(math.ceil(allf/self.max_seq_len))
-        self.example_counts = example_counts
-
-    
-    def process_split(self, split, indexers) -> Iterable[Type[Instance]]:
-        """Process a language modeling split by indexing and creating fields.
-        Args:
-            split: (list) a single list of sentences
-            indexers: (Indexer object) indexer to index input words
-        """
-        def _make_instance(sent):
-            ''' Forward targs adds <s> as a target for input </s>
-            and bwd targs adds </s> as a target for input <s>
-            to avoid issues with needing to strip extra tokens
-            in the input for each direction '''
-            d = {}
-            #import pdb;pdb.set_trace()
-            d["input"] = sentence_to_text_field(sent[:-1], indexers)
-            d["targs"] = sentence_to_text_field(sent[1:], self.target_indexer)
-            d["targs_b"] = sentence_to_text_field([sent[-1]] + sent[:-2], self.target_indexer)
-            return Instance(d)
-        for sent in split:
-            yield _make_instance(sent)
-    def load_data(self, path):
-        """Loading data file and tokenizing the text
-        Args:
-            path: (str) data file path
-        """
-        seq_len = self.max_seq_len
-        tokens = []
-        targ_map = {'neutral': 0, 'entailment': 1, 'contradiction': 2}
-        data = load_tsv(os.path.join(path), 1000, skip_rows=1, s1_idx=8, s2_idx=9, targ_idx=11, targ_map=targ_map)
-        rows=[]
-        tokens=[]
-        for x, y in zip(data[0], data[1]):
-            tokens += x[1:-1]+["<EOS>"]+y[1:-1]+["<EOS>"]
-        for i in range(0, len(tokens), seq_len):
-            yield tokens[i:i+seq_len]
-
-
-
