+++ src/utils/locked_dropout.py
@@ -2,6 +2,7 @@ import torch
 import torch.nn as nn
 from torch.autograd import Variable
 
+
 class LockedDropout(nn.Module):
     def __init__(self):
         super().__init__()
